### 2024-10-08 #7332 @crvernon
:wave: @toruseo - Could you please provide a branch that has both the code of your repo and the paper components needed for JOSS within the same branch?  This will help our editorialbot to find what it needs.
