+++ paper/paper.md
@@ -1,0 +1,226 @@
+---
+title: 'UXsim: lightweight mesoscopic traffic flow simulator in pure Python'
+tags:
+  - Python
+  - transportation
+  - traffic flow
+  - civil engineering
+  - simulation
+authors:
+  - name: Toru Seo
+    orcid: 0000-0001-9767-7218
+    affiliation: "1"
+affiliations:
+ - name: Department of Civil and Environmental Engineering, Institute of Science Tokyo, Japan
+   index: 1
+date: 2024-10-07
+bibliography: paper.bib
+---
+
+# Summary
+
+Car traffic plays an essential role in today's cities.
+However, it faces severe issues such as time loss due to congestion, accidents, and environmental pollution.
+Proper traffic planning and control are necessary to solve these issues.
+Traffic flow simulators are powerful engineering tools for such purposes.
+By using traffic flow simulators, one can test and optimize traffic plans or control strategies before implementing them in the real world.
+Furthermore, traffic flow simulators can be used for pure academic or scientific experiments to investigate the complex physics of traffic flow.
+
+*UXsim* is a pure Python package for traffic flow simulation that is suitable for simulating a wide range of scenarios, from simple toy networks to large-scale (e.g., city-wide) traffic phenomena.
+It computes dynamic traffic flow in arbitrary road networks with arbitrary travel demands.
+For example, it can simulate how traffic jams evolve in a city, how people reroute to avoid them, and how traffic management schemes can improve the situation.
+
+# Statement of need
+
+In general, traffic flow simulators are divided into two major categories: microscopic and macroscopic.
+The former focuses on detailed individual vehicle behaviors, whereas the latter focuses on the collective behavior of large-scale (e.g., city-scale) traffic.
+Mesoscopic traffic flow simulators, which is sometimes categorized into macroscopic ones, are a mixture of the two; while they describe individual vehicle behaviors to some extent, their primary intention is to simulate the collective behavior of large-scale traffic.
+Mesoscopic simulators are particularly useful for modeling large-scale traffic management and operations, such as congestion pricing, ridesharing, and automated vehicle fleet management, which are increasingly prominent these days.
+Several microscopic traffic flow simulators are published as open source software, such as SUMO [@sumo].
+However, to the author's knowledge, the availability of mesoscopic and macroscopic simulators is limited.
+
+*UXsim* is a mesoscopic (or macroscopic depending on the technical terminology) traffic flow simulator.
+It is particularly designed to be simple, lightweight, and easy to use, thanks to the simplicity of mesoscopic traffic flow models.
+Therefore, it can be readily used by users to simulate small-scale to large-scale traffic flow.
+It also includes built-in analysis and visualization capabilities for simulation results.
+
+A unique feature of *UXsim* is that it is written in pure Python.
+Except for common components in *NumPy* and *SciPy*, no pre-compiled binary is used in *UXsim*.
+Therefore, it can be easily and deeply integrated with other Python-based frameworks, such as *PyTorch* for deep reinforcement learning-based traffic control.
+Users with Python knowledge can customize the code of *UXsim* to make it suitable for their own purposes.
+This makes *UXsim* a powerful tool for state-of-the-art research.
+Nevertheless, the computation speed is not too slow, leveraging mesoscopic traffic flow simulation techniques and the *NumPy* and *SciPy* packages.
+
+*UXsim* can be used by any user who has basic Python skills and an interest in traffic.
[…]
