+++ paper/paper.bib
@@ -1,0 +1,195 @@
+@article{microsofttrace,
+  title={Trace is the New {AutoDiff}--Unlocking Efficient Optimization of Computational Workflows},
+  author={Cheng, Ching-An and Nie, Allen and Swaminathan, Adith},
+  journal={arXiv preprint arXiv:2406.16218},
+  doi={10.48550/arXiv.2406.16218},
+  year={2024}
+}
+
+@article{iizuka,
+  title={Passenger-oriented Distributed Traffic Signal Control Using Dynamic Programming with Vehicle Queue Spillback and Waiting Time Constraints},
+  author={Iizuka, Kanae and Seo, Toru},
+  journal={International Journal of Intelligent Transportation Systems Research},
+  year={2024},
+  doi={10.1007/s13177-024-00418-z},
+  publisher={Springer}
+}
+
+@inproceedings{sumo,
+  title={Microscopic traffic simulation using {SUMO}},
+  author={Lopez, Pablo Alvarez and Behrisch, Michael and Bieker-Walz, Laura and Erdmann, Jakob and Fl{\"o}tter{\"o}d, Yun-Pang and Hilbrich, Robert and L{\"u}cken, Leonhard and Rummel, Johannes and Wagner, Peter and Wie{\ss}ner, Evamarie},
+  booktitle={2018 21st international conference on intelligent transportation systems},
+  pages={2575--2582},
+  year={2018},
+  organization={IEEE}
+}
+
+@Article{movsim,
+  author  = {Treiber, Martin and Kesting, Ame},
+  journal = {IEEE Intelligent Transportation Systems Magazine},
+  title   = {An open-source microscopic traffic simulator},
+  year    = {2010},
+  number  = {3},
+  pages   = {6--13},
+  volume  = {2},
+  doi     = {10.1109/MITS.2010.939208},
+}
+
+@misc{abstreet,
+  author = {Dustin Carlino and Yuwen Li and Michael Kirk},
+  title = {{A/B Street}},
+  year = {2018},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/a-b-street/abstreet}
+}
+
+@article{cityflow,
+  title={{Cityflow}: A multi-agent reinforcement learning environment for large scale city traffic scenario},
+  author={Zhang, Huichu and Feng, Siyuan and Liu, Chang and Ding, Yaoyao and Zhu, Yichen and Zhou, Zihan and Zhang, Weinan and Yu, Yong and Jin, Haiming and Li, Zhenhui},
+  journal={arXiv preprint arXiv:1905.05217},
+  doi={10.48550/arXiv.2406.16218},
+  year={2019}
+}
+
+@Article{dyntapy,
+  author  = {Ortmann, Paul and Tamp{\`e}re, Chris M. J.},
+  journal = {Journal of Open Source Software},
+  title   = {{dyntapy}: dynamic and static traffic assignment in {Python}},
+  year    = {2022},
+  issn    = {2475-9066},
+  number  = {77},
+  pages   = {4593},
+  volume  = {7},
+  doi     = {10.21105/joss.04593},
+  file    = {:Ortmann2022dta.pdf:PDF},
+}
+
+@misc{dtalite,
+  author = {{DTALite Team}},
+  title = {{DTALite}},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/asu-trans-ai-lab/DTALite}
+}
+
+@Article{Zhou2014dtalite,
+  author    = {Zhou, Xuesong and Taylor, Jeffrey},
+  journal   = {Cogent Engineering},
+  title     = {{DTALite}: A queue-based mesoscopic traffic simulator for fast model evaluation and calibration},
+  year      = {2014},
+  number    = {1},
[…]
