### 2024-10-09 #7332 @toruseo
@crvernon Sorry for my misunderstanding. Now branch `JOSS` has both of the paper and software code.
