### 2024-10-10 #7332 @Nitnelav
> Hi @Nitnelav, I'll keep that in mind when I get an editor assigned to this submission. How did you hear about this review by the way?

Cool, thanks !

I had an eye on the joss-reviews issues for some time, looking for one I could volunteer to review. I was hipped about JOSS at an "open source in acoustics" workshop and I found the open submission and reviewing process on github very interesting. Then we (the [UMRAE](https://www.umrae.fr/en/) team) decided to submit our [NoiseModelling](https://github.com/Universite-Gustave-Eiffel/NoiseModelling) software to JOSS (work is in progress). And so I thought I'd participate to the reviewing process to see how it goes !
