### 2024-10-10 #7332 @xuanxu
The error in the PDF generation comes from [this date](https://github.com/toruseo/UXsim/blob/JOSS/paper/paper.md?plain=1#L16) in the paper's metadata not following [the expected syntax ](https://joss.readthedocs.io/en/latest/example_paper.html#example-paper).
