### 2024-10-10 388f5b3
fix date format
https://github.com/openjournals/joss-reviews/issues/7332#issuecomment-2404359129
