### 2024-11-05 #7332 @toruseo
@kanishkan91 Thank you very much for handling my paper.

Do I need to do something to advance the review process? Or should I just wait? Sorry for interrupting you. I understand that the editorial process take some time, but I am not familiar with this submission system.
