### 2024-11-08 #7332 @toruseo
Thank you.

Regarding potential reviewers, [Nitnelav](https://github.com/Nitnelav) kindly volunteered as a reviewer (pls see https://github.com/openjournals/joss-reviews/issues/7332#issuecomment-2404335660 for his motivation).
The authors of [the related papers](https://github.com/openjournals/joss-reviews/issues/7332#issuecomment-2404475175) and [the paper cited by my manuscript](https://joss.theoj.org/papers/10.21105/joss.04593) would be also great.
