### 2024-11-12 c255c70
Add DUE example and related functions
