+++ uxsim/uxsim.py
@@ -182,10 +182,10 @@ class Node:
 
                     #consider the link preferences
                     outlinks = list(s.outlinks.values())
-                    if set(outlinks) & set(veh.links_prefer): 
-                        outlinks = sorted(set(outlinks) & set(veh.links_prefer))
+                    if set(outlinks) & set(veh.links_prefer):
+                        outlinks = sorted(set(outlinks) & set(veh.links_prefer), key=lambda l:l.name)
                     if set(outlinks) & set(veh.links_avoid):
-                        outlinks = sorted(set(outlinks) - set(veh.links_avoid))
+                        outlinks = sorted(set(outlinks) - set(veh.links_avoid), key=lambda l:l.name)
                     
                     preference = np.array([veh.route_pref[l.id] for l in outlinks], dtype=float)
                     if s.W.hard_deterministic_mode == False:
@@ -1105,6 +1105,13 @@ class Vehicle:
         if s.x_next > s.link.length:
             s.move_remain = s.x_next - s.link.length
             s.x_next = s.link.length
+    
+    def enforce_route(s, route):
+        """
+        Enforce the vehicle to use the specified route. The route should connect the origin to the destination. TODO: add consistency check
+        """
+        s.links_prefer = [s.W.get_link(l) for l in route]
+        s.links_avoid = [s.W.get_link(l) for l in s.W.LINKS if l not in s.links_prefer]
 
     def route_pref_update(s, weight=1):
         """
@@ -1156,9 +1163,9 @@ class Vehicle:
 
                 #if links_prefer is given and available at the node, select only from the links in the list. if links_avoid is given, select links not in the list.
                 if set(outlinks) & set(s.links_prefer):
-                    outlinks = sorted(set(outlinks) & set(s.links_prefer))
+                    outlinks = sorted(set(outlinks) & set(s.links_prefer), key=lambda l:l.name)
                 if set(outlinks) & set(s.links_avoid):
-                    outlinks = sorted(set(outlinks) - set(s.links_avoid))
+                    outlinks = sorted(set(outlinks) - set(s.links_avoid), key=lambda l:l.name)
 
                 preference = np.array([s.route_pref[l.id] for l in outlinks], dtype=float)
                 if s.W.hard_deterministic_mode == False:
@@ -1232,6 +1239,13 @@ class Vehicle:
     def traveled_route(s):
         """
         Returns the route this vehicle traveled.
+
+        Returns
+        -------
+        Route
+            The route this vehicle traveled.
+        ts
+            The time at which the vehicle entered each link. The last element is the time the vehicle reached the destination.
         """
         link_old = -1
         t = -1
@@ -1944,7 +1958,11 @@ class World:
         ds = W.rng.choice(dests, size=size)
         
         for t, o, d, in zip(ts, os, ds):
-            W.addVehicle(o, d, t, attribute=attribute, direct_call=False)
+            d2 = d
+            if o == d:
+                d2 = W.rng.choice([dd for dd in dests if dd!=d])
+            W.addVehicle(o, d2, t, attribute=attribute, direct_call=False)
+
 
     @demand_info_record
     def adddemand_nodes2nodes2(W, origs, dests, t_start, t_end, flow=-1, volume=-1, attribute=None, direct_call=True):
@@ -1991,7 +2009,10 @@ class World:
         ds = W.rng.choice(dests, size=size)
         
         for t, o, d, in zip(ts, os, ds):
-            W.addVehicle(o, d, t, attribute=attribute, direct_call=False)
+            d2 = d
+            if o == d:
+                d2 = W.rng.choice([dd for dd in dests if dd!=d])
[…]
