### 2024-11-13 ebf8ac4
fix bug in DUE example
