+++ demos_and_examples/example_25en_dynamic_user_equilibrium.py
@@ -1,22 +1,31 @@
 """
-This code is to compute near dynamic user equilibirum state as a steady state (a proxy for Nash equilibrium) of day-to-day dynamical routing game.
-Work in progress.
+This code is to compute a near dynamic user equilibirum state as a steady state of day-to-day dynamical routing game.
+
+Specifically, on day `i`, vehicles choose their route based on actual travel time on day `i-1` with the same departure time.
+If there are shorter travel time route, they will change with probability `swap_prob`.
+This process is repeated until `max_iter` day.
+It is expected that this process eventually reach a steady state.
+Due to the problem complexity, it does not necessarily reach or converge to Nash equilibrium or any other stationary points.
+However, in the literature, it is argued that the steady state can be considered as a reasonable proxy for Nash equilibrium or dynamic equilibrium state.
+There are some theoretical background for it; but intuitively speaking, the steady state can be considered as a realistic state that people's rational behavior will reach.
 """
 
 from pylab import *
 import uxsim
-from uxsim.Utilities import enumerate_k_shortest_routes, generate_grid_network
+from uxsim.Utilities import enumerate_k_shortest_routes
 import random
 from collections import defaultdict
 
 # scenario definition
+seed = None
 W_orig = uxsim.World(
     name="",
     deltan=10,
     tmax=4000,
     print_mode=0, save_mode=1, show_mode=1,
-    random_seed=None
+    random_seed=seed
 )
+random.seed(seed)
 
 n_nodes = 4
 imax = n_nodes
@@ -38,13 +47,12 @@ for i in range(imax):
         if j != 0:
             links[i,j,i,j-1] = W_orig.addLink(f"l{(i,j,i,j-1)}", nodes[i,j], nodes[i,j-1], length=1000)
 
-
 W_orig.adddemand_nodes2nodes2(nodes.values(), nodes.values(), 0, 3000, volume=20000)
 
 W_orig.print_scenario_stats()
 
 # enumerate routes for each OD pair
-n_routes_per_od = 3
+n_routes_per_od = 4
 
 dict_od_to_vehid = defaultdict(lambda: [])
 for key, veh in W_orig.VEHICLES.items():
@@ -61,13 +69,11 @@ for o,d in dict_od_to_vehid.keys():
 print(f"number of OD pairs: {len(dict_od_to_routes.keys())}, number of routes: {sum([len(val) for val in dict_od_to_routes.values()])}")
 
 # day-to-day dynamics
-print("starting day-to-day dynamics...")
 ttts = []
 n_swaps = []
 potential_swaps = []
 total_t_gaps = []
 swap_prob = 0.05
-swap_threth = 1.00
 max_iter = 100
 
 for i in range(max_iter):
@@ -75,7 +81,8 @@ for i in range(max_iter):
 
     if i != 0:
         for key in W.VEHICLES:
-            W.VEHICLES[key].enforce_route(routes_specified[key])
+            if key in routes_specified:
+                W.VEHICLES[key].enforce_route(routes_specified[key])
     
     route_set = defaultdict(lambda: []) #routes[o,d] = [Route, Route, ...]
     for o,d in dict_od_to_routes.keys():
@@ -89,37 +96,39 @@ for i in range(max_iter):
     W.analyzer.print_simple_stats()
 
     # route swap
-
     routes_specified = {}
     n_swap = 0
     total_t_gap = 0
     potential_n_swap = 0
     for key,veh in W.VEHICLES.items():
+        if veh.state != "end":
+            continue
+
         o = veh.orig.name
         d = veh.dest.name
         r, ts = veh.traveled_route()
         travel_time = ts[-1]-ts[0]
 
-        should_change = False
-        should_change_to = None
+        flag_route_changed = False
[…]
