### 2024-11-14 068347f
improve `log_t_link`
