### 2024-11-21 a450e92
update the logic of DSO example22
