+++ demos_and_examples/example_22en_routing_optimization.py
@@ -1,35 +1,87 @@
-from deap import base, creator, tools, algorithms
 from pylab import *
-from uxsim import *
-from uxsim import Utilities
+import uxsim
+from uxsim.Utilities import enumerate_k_shortest_routes, generate_grid_network
 import random
+from collections import defaultdict
+import random
+
+def initialize_population(pop_size, length, n_max):
+    """Initialize a population with random genomes."""
+    population = []
+    for _ in range(pop_size):
+        genome = [random.randint(0, n_max) for _ in range(length)]
+        population.append(genome)
+    return population
+
+def tournament_selection(population, fitnesses, k=3):
+    """Select individuals using tournament selection."""
+    selected = []
+    for _ in range(len(population)):
+        aspirants = random.sample(list(zip(population, fitnesses)), k)
+        selected.append(max(aspirants, key=lambda x: x[1])[0])
+    return selected
+
+def crossover(parent1, parent2, n_points=1):
+    """Perform n-point crossover between two parents."""
+    if n_points >= len(parent1):
+        raise ValueError("Number of crossover points must be less than genome length.")
+    
+    points = sorted(random.sample(range(1, len(parent1)), n_points))
+    child1, child2 = [], []
+    last_point = 0
+    for i, point in enumerate(points + [len(parent1)]):
+        if i % 2 == 0:
+            child1.extend(parent1[last_point:point])
+            child2.extend(parent2[last_point:point])
+        else:
+            child1.extend(parent2[last_point:point])
+            child2.extend(parent1[last_point:point])
+        last_point = point
+    return child1, child2
+
+def mutate(genome, mutation_rate, n_max):
+    """Mutate a genome with a given mutation rate."""
+    mutated_genome = []
+    for gene in genome:
+        if random.random() < mutation_rate:
+            mutated_gene = random.randint(0, n_max)
+            mutated_genome.append(mutated_gene)
+        else:
+            mutated_genome.append(gene)
+    return mutated_genome
+
+
 
 ##############################################################
 # Define UXsim World
-W = World(
-   name="", 
-   deltan=5, 
-   tmax=1200,
-   print_mode=0, save_mode=1, show_mode=1, #print is off, otherwise it will be very verbose during the genetic algorithm
-   random_seed=None,
-   duo_update_time=60
-)
+def create_World():
+    W = uxsim.World(
+        name="", 
+        deltan=5, 
+        tmax=1200,
+        print_mode=0, save_mode=1, show_mode=1, #print is off, otherwise it will be very verbose during the genetic algorithm
+        duo_update_time=60, 
+        vehicle_logging_timestep_interval=-1, 
+        random_seed=42
+    )
 
-## generate grid network
-Utilities.generate_grid_network(W, 3, 3, length=1000)
-W.show_network()
+    ## generate grid network
+    generate_grid_network(W, 3, 3, length=1000)
+    # W.show_network()
 
-## set demand
-od_pairs = [
-   ("n(0, 0)", "n(2, 2)"),
-   ("n(2, 0)", "n(0, 2)"),
-   ("n(0, 2)", "n(2, 0)"),
-   ("n(2, 2)", "n(0, 0)"),
-]
-for od_pair in od_pairs:
-   W.adddemand(od_pair[0], od_pair[1], 0, 500, 1)
+    ## set demand
+    od_pairs = [
+        ("n(0, 0)", "n(2, 2)"),
+        ("n(2, 0)", "n(0, 2)"),
+        ("n(0, 2)", "n(2, 0)"),
[…]
