+++ uxsim/DTAsolvers/DTAsolvers.py
@@ -1,0 +1,178 @@
+"""
+Submodule for Dynamic Traffic Assginment solvers
+"""
+import random
+import matplotlib.pyplot as plt
+import numpy as np
+from collections import defaultdict
+from ..Utilities import enumerate_k_shortest_routes
+
+def solve_DUE(func_World, max_iter=50, n_routes_per_od=4, swap_prob=0.05, print_info=True, plot_res=False):
+    """
+    Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics
+
+    Parameters
+    ----------
+    func_World : function
+        function that returns a World object with nodes, links, and demand specifications
+    max_iter : int
+        maximum number of iterations
+    n_routes_per_od : int
+        number of routes to enumerate for each OD pair
+    swap_prob : float
+        probability of route swap
+    print_info : bool
+        whether to print the information
+    plot_res : bool
+        whether to plot the results
+
+    Returns
+    -------
+    W : World
+        World object with quasi DUE solution (if properly converged)
+    
+    Notes
+    -----
+        This function computes a near dynamic user equilibirum state as a steady state of day-to-day dynamical routing game.
+
+        Specifically, on day `i`, vehicles choose their route based on actual travel time on day `i-1` with the same departure time.
+        If there are shorter travel time route, they will change with probability `swap_prob`.
+        This process is repeated until `max_iter` day.
+        It is expected that this process eventually reach a steady state.
+        Due to the problem complexity, it does not necessarily reach or converge to Nash equilibrium or any other stationary points.
+        However, in the literature, it is argued that the steady state can be considered as a reasonable proxy for Nash equilibrium or dynamic equilibrium state.
+        There are some theoretical background for it; but intuitively speaking, the steady state can be considered as a realistic state that people's rational behavior will reach.
+    """
+
+    W_orig = func_World()
+    if print_info:
+        W_orig.print_scenario_stats()
+
+    # enumerate routes for each OD pair
+    n_routes_per_od = n_routes_per_od
+
+    dict_od_to_vehid = defaultdict(lambda: [])
+    for key, veh in W_orig.VEHICLES.items():
+        o = veh.orig.name
+        d = veh.dest.name
+        dict_od_to_vehid[o,d].append(key)
+
+    dict_od_to_routes = {}
+    for o,d in dict_od_to_vehid.keys():
+        routes = enumerate_k_shortest_routes(W_orig, o, d, k=n_routes_per_od) #TODO: to be replaced with many-to-many shortest path search
+        dict_od_to_routes[o,d] = routes
+
+    if print_info:
+        print(f"number of OD pairs: {len(dict_od_to_routes.keys())}, number of routes: {sum([len(val) for val in dict_od_to_routes.values()])}")
+
+    # day-to-day dynamics
+    ttts = []
+    n_swaps = []
+    potential_swaps = []
+    total_t_gaps = []
+    swap_prob = swap_prob
+    max_iter = max_iter #50 or 100 is better
+
+    if print_info:
+        print("solving DUE...")
+    for i in range(max_iter):
+        W = func_World()
+
+        if i != 0:
+            for key in W.VEHICLES:
+                if key in routes_specified:
+                    W.VEHICLES[key].enforce_route(routes_specified[key])
+        
+        route_set = defaultdict(lambda: []) #routes[o,d] = [Route, Route, ...]
+        for o,d in dict_od_to_routes.keys():
+            for r in dict_od_to_routes[o,d]:
+                route_set[o,d].append(W.defRoute(r))
+
+        # simulation
+        W.exec_simulation()
+
[…]
