### 2024-11-25 cc062b1
disable OSMnx tests
