### 2024-11-26 2479697
add DSO solvers (WIP)
