+++ uxsim/DTAsolvers/DTAsolvers.py
@@ -6,10 +6,11 @@ import matplotlib.pyplot as plt
 import numpy as np
 from collections import defaultdict
 from ..Utilities import enumerate_k_shortest_routes
+from ..utils import *
 
 def solve_DUE(func_World, max_iter=50, n_routes_per_od=4, swap_prob=0.05, print_info=True, plot_res=False):
     """
-    Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics
+    Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics. WIP
 
     Parameters
     ----------
@@ -42,6 +43,10 @@ def solve_DUE(func_World, max_iter=50, n_routes_per_od=4, swap_prob=0.05, print_
         Due to the problem complexity, it does not necessarily reach or converge to Nash equilibrium or any other stationary points.
         However, in the literature, it is argued that the steady state can be considered as a reasonable proxy for Nash equilibrium or dynamic equilibrium state.
         There are some theoretical background for it; but intuitively speaking, the steady state can be considered as a realistic state that people's rational behavior will reach.
+
+        This method is based on the following literature:
+        Ishihara, M., & Iryo, T. (2015). Dynamic Traffic Assignment by Markov Chain. Journal of Japan Society of Civil Engineers, Ser. D3 (Infrastructure Planning and Management), 71(5), I_503-I_509. (in Japanese). https://doi.org/10.2208/jscejipm.71.I_503
+        Iryo, T., Urata, J., & Kawase, R. (2024). Traffic Flow Simulator and Travel Demand Simulators for Assessing Congestion on Roads After a Major Earthquake. In APPLICATION OF HIGH-PERFORMANCE COMPUTING TO EARTHQUAKE-RELATED PROBLEMS (pp. 413-447). https://doi.org/10.1142/9781800614635_0007
     """
 
     W_orig = func_World()
@@ -77,6 +82,8 @@ def solve_DUE(func_World, max_iter=50, n_routes_per_od=4, swap_prob=0.05, print_
         print("solving DUE...")
     for i in range(max_iter):
         W = func_World()
+        if i != max_iter-1:
+            W.vehicle_logging_timestep_interval = -1
 
         if i != 0:
             for key in W.VEHICLES:
@@ -176,3 +183,389 @@ def solve_DUE(func_World, max_iter=50, n_routes_per_od=4, swap_prob=0.05, print_
         plt.show()
     
     return W
+
+def solve_DSO(func_World, max_iter=300, n_routes_per_od=4, beta_coef=100, beta_coef2=5000, print_info=True, print_info_detailed=False, plot_res=False, initial_solution_World=None):
+    """
+    Solve Dynamic System Optimum (DSO) problem. WIP
+
+    Parameters
+    ----------
+    func_World : function
+        function that returns a World object with nodes, links, and demand specifications
+    max_iter : int
+        maximum number of iterations
+    n_routes_per_od : int
+        number of routes to enumerate for each OD pair
+    beta_coef : float
+        coefficient for logit response dynamics
+    beta_coef2 : float
+        coefficient for logit response dynamics
+    print_info : bool
+        whether to print the information
+    print_info_detailed : bool
+        whether to print the detailed information
+    plot_res : bool
+        whether to plot the results
+    initial_solution_World : World
+        initial solution (starting point) for the optimization algorithm. It must have the same structure as the output of func_World, and its simulation has been already executed. Recommended example: `W_init = solve_DUE(func_World); W = solve_DSO(func_World, initial_solution_World=W_init)`
+
+    Returns
+    -------
+    W : World
+        World object with near DSO solution (if properly converged)
+    
+    Notes
+    -----
[…]
