+++ uxsim/DTAsolvers/DTAsolvers.py
@@ -5,10 +5,270 @@ import random
 import matplotlib.pyplot as plt
 import numpy as np
 from collections import defaultdict
-from ..Utilities import enumerate_k_shortest_routes
+from ..Utilities import enumerate_k_shortest_routes, enumerate_k_random_routes
 from ..utils import *
 
-def solve_DUE(func_World, max_iter=50, n_routes_per_od=4, swap_prob=0.05, print_info=True, plot_res=False):
+class Solver_DUE:
+    def __init__(s, func_World):
+        """
+        Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics. WIP
+
+        Parameters
+        ----------
+        func_World : function
+            function that returns a World object with nodes, links, and demand specifications
+            
+        Notes
+        -----
+            This function computes a near dynamic user equilibirum state as a steady state of day-to-day dynamical routing game.
+
+            Specifically, on day `i`, vehicles choose their route based on actual travel time on day `i-1` with the same departure time.
+            If there are shorter travel time route, they will change with probability `swap_prob`.
+            This process is repeated until `max_iter` day.
+            It is expected that this process eventually reach a steady state.
+            Due to the problem complexity, it does not necessarily reach or converge to Nash equilibrium or any other stationary points.
+            However, in the literature, it is argued that the steady state can be considered as a reasonable proxy for Nash equilibrium or dynamic equilibrium state.
+            There are some theoretical background for it; but intuitively speaking, the steady state can be considered as a realistic state that people's rational behavior will reach.
+
+            This method is based on the following literature:
+            Ishihara, M., & Iryo, T. (2015). Dynamic Traffic Assignment by Markov Chain. Journal of Japan Society of Civil Engineers, Ser. D3 (Infrastructure Planning and Management), 71(5), I_503-I_509. (in Japanese). https://doi.org/10.2208/jscejipm.71.I_503
+            Iryo, T., Urata, J., & Kawase, R. (2024). Traffic Flow Simulator and Travel Demand Simulators for Assessing Congestion on Roads After a Major Earthquake. In APPLICATION OF HIGH-PERFORMANCE COMPUTING TO EARTHQUAKE-RELATED PROBLEMS (pp. 413-447). https://doi.org/10.1142/9781800614635_0007
+        """
+        s.func_World = func_World
+        s.W_sol = None
+        s.dfs_link = []
+    
+    def solve(s, max_iter, n_routes_per_od=4, swap_prob=0.05, print_info=True):
+        """
+        Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics. WIP
+
+        Parameters
+        max_iter : int
+            maximum number of iterations
+        n_routes_per_od : int
+            number of routes to enumerate for each OD pair
+        swap_prob : float
+            probability of route swap
+        print_info : bool
+            whether to print the information
+
+        Returns
+        -------
+        W : World
+            World object with quasi DUE solution (if properly converged)
+        """
+
+        W_orig = s.func_World()
+        if print_info:
+            W_orig.print_scenario_stats()
+
+        # enumerate routes for each OD pair
+        n_routes_per_od = n_routes_per_od
+
+        dict_od_to_vehid = defaultdict(lambda: [])
[…]
