+++ tests/test_examples.py
@@ -13,7 +13,6 @@ nontest_files = [
     "example_14en_multiple_signals_Deep_Reinforcement_Learning_pytorch.py", #takes too much time. DQN can be tested by example_12en_* as well.
     "example_17en_result_GUI_viewer_sioux_falls.py", #GUI viewer is not testable in Github Actions
     "example_18en_result_GUI_viewer_grid.py", #GUI viewer is not testable in Github Actions
-    "example_16en_import_from_OpenStreetMap.py", #inconsist with OSMnx 2.0.0. TODO: fix
 ]
 
 # Dynamically generate test cases for each example script
