+++ tests/test_other_functions.py
@@ -77,35 +77,34 @@ def test_analyzer():
 
     assert True
 
-#inconsist with OSMnx 2.0.0. TODO: fix
-# @pytest.mark.filterwarnings("ignore::UserWarning")
-# def test_osm_import():
-#     from uxsim.OSMImporter import OSMImporter
+@pytest.mark.filterwarnings("ignore::UserWarning")
+def test_osm_import():
+    from uxsim.OSMImporter import OSMImporter
 
-#     W = World(
-#         name="",
-#         deltan=5,
-#         tmax=7200,
-#         print_mode=1, save_mode=1, show_mode=0, 
-#         random_seed=0
-#     )
+    W = World(
+        name="",
+        deltan=5,
+        tmax=7200,
+        print_mode=1, save_mode=1, show_mode=0, 
+        random_seed=0
+    )
 
-#     #Tokyo highway
-#     nodes, links = OSMImporter.import_osm_data(bbox=(35.817, 35.570, 139.881, 139.583), custom_filter='["highway"~"motorway"]')
-#     nodes, links = OSMImporter.osm_network_postprocessing(nodes, links, node_merge_threshold=0.005, node_merge_iteration=5, enforce_bidirectional=True)  # merge threshold distance: 0.005 degree ~= 500 m. `enforce_bidirectional` makes all links bidirectional, so that network is not fragmented (but the original network topology is not preserved rigorously).
-#     OSMImporter.osm_network_visualize(nodes, links, show_link_name=0, show_mode=0, save_mode=0)
-#     OSMImporter.osm_network_to_World(W, nodes, links, default_jam_density=0.2, coef_degree_to_meter=111000)
-#     W.show_network()
-    
-#     # set demand to central Tokyo from surroundings
-#     W.adddemand_area2area(139.70, 35.60, 0, 139.75, 35.68, 0.05, 0, 3600, volume=5000)
-#     W.adddemand_area2area(139.65, 35.70, 0, 139.75, 35.68, 0.05, 0, 3600, volume=5000)
-#     W.adddemand_area2area(139.75, 35.75, 0, 139.75, 35.68, 0.05, 0, 3600, volume=5000)
-#     W.adddemand_area2area(139.85, 35.70, 0, 139.75, 35.68, 0.05, 0, 3600, volume=5000)
+    #Tokyo highway
+    nodes, links = OSMImporter.import_osm_data(bbox=(35.817, 35.570, 139.881, 139.583), custom_filter='["highway"~"motorway"]')
+    nodes, links = OSMImporter.osm_network_postprocessing(nodes, links, node_merge_threshold=0.005, node_merge_iteration=5, enforce_bidirectional=True)  # merge threshold distance: 0.005 degree ~= 500 m. `enforce_bidirectional` makes all links bidirectional, so that network is not fragmented (but the original network topology is not preserved rigorously).
+    OSMImporter.osm_network_visualize(nodes, links, show_link_name=0, show_mode=0, save_mode=0)
+    OSMImporter.osm_network_to_World(W, nodes, links, default_jam_density=0.2, coef_degree_to_meter=111000)
+    W.show_network()
+  
+    # set demand to central Tokyo from surroundings
+    W.adddemand_area2area(139.70, 35.60, 0, 139.75, 35.68, 0.05, 0, 3600, volume=5000)
+    W.adddemand_area2area(139.65, 35.70, 0, 139.75, 35.68, 0.05, 0, 3600, volume=5000)
+    W.adddemand_area2area(139.75, 35.75, 0, 139.75, 35.68, 0.05, 0, 3600, volume=5000)
[…]
