### 2024-12-25 #7332 @kanishkan91
@Nitnelav would you be interested in reviewing this paper and package for JOSS?
