### 2024-12-25 #7332 @kanishkan91
@maximelenormand, based on your previous paper, reviews, would you be interested in reviewing this paper for JOSS?
