### 2024-12-25 #7332 @kanishkan91
@fxjung would you be interested in reviewing this paper for JOSS? It seems like a good fit given previous papers youve authored
