### 2024-12-25 3fe3c2a
update DTAsolver: full class
