+++ uxsim/DTAsolvers/DTAsolvers.py
@@ -8,7 +8,9 @@ from collections import defaultdict
 from ..Utilities import enumerate_k_shortest_routes, enumerate_k_random_routes
 from ..utils import *
 
-class Solver_DUE:
+import warnings
+
+class SolverDUE:
     def __init__(s, func_World):
         """
         Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics. WIP
@@ -33,6 +35,7 @@ class Solver_DUE:
             This method is based on the following literature:
             Ishihara, M., & Iryo, T. (2015). Dynamic Traffic Assignment by Markov Chain. Journal of Japan Society of Civil Engineers, Ser. D3 (Infrastructure Planning and Management), 71(5), I_503-I_509. (in Japanese). https://doi.org/10.2208/jscejipm.71.I_503
             Iryo, T., Urata, J., & Kawase, R. (2024). Traffic Flow Simulator and Travel Demand Simulators for Assessing Congestion on Roads After a Major Earthquake. In APPLICATION OF HIGH-PERFORMANCE COMPUTING TO EARTHQUAKE-RELATED PROBLEMS (pp. 413-447). https://doi.org/10.1142/9781800614635_0007
+            Iryo, T., Watling, D., & Hazelton, M. (2024). Estimating Markov Chain Mixing Times: Convergence Rate Towards Equilibrium of a Stochastic Process Traffic Assignment Model. Transportation Science. https://doi.org/10.1287/trsc.2024.0523
         """
         s.func_World = func_World
         s.W_sol = None
@@ -43,6 +46,7 @@ class Solver_DUE:
         Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics. WIP
 
         Parameters
+        ----------
         max_iter : int
             maximum number of iterations
         n_routes_per_od : int
@@ -265,7 +269,575 @@ class Solver_DUE:
         plt.legend()
         plt.show()
 
+class SolverDSO:
+    def __init__(s, func_World):
+        """
+        Solve Dynamic System Optimum (DSO) problem. WIP
+
+        Parameters
+        ----------
+        func_World : function
+            function that returns a World object with nodes, links, and demand specifications
             
+        Notes
+        -----
+            This is based on an algorithm whose stochastic convergence is theoretically guaranteed. However, the convergence is very slow. It is difficult to obtain accurate solution within a practical amount of time if the network was not small. 
+            
+            This method is based on the following literature:
+            Satsukawa, K., Wada, K., & Watling, D. (2022). Dynamic system optimal traffic assignment with atomic users: Convergence and stability. Transportation Research Part B: Methodological, 155, 188-209. https://doi.org/10.1016/j.trb.2021.11.001
+        """
+        s.func_World = func_World
+        s.W_sol = None
+        s.dfs_link = []
+    
+    def solve(s, max_iter, n_routes_per_od=4, beta_coef=100, beta_coef2=5000, print_info=True, print_info_detailed=False, plot_res=False, initial_solution_World=None):
+        """
+        Solve Dynamic System Optimum (DSO) problem. WIP
+
+        Parameters
+        ----------
+        max_iter : int
+            maximum number of iterations
+        n_routes_per_od : int
+            number of routes to enumerate for each OD pair
+        beta_coef : float
+            coefficient for logit response dynamics
+        beta_coef2 : float
+            coefficient for logit response dynamics
+        print_info : bool
+            whether to print the information
+        print_info_detailed : bool
+            whether to print the detailed information
+        plot_res : bool
+            whether to plot the results
+        initial_solution_World : World
[…]
