+++ pyproject.toml
@@ -28,7 +28,7 @@ dependencies = [
     "scipy>=1.9.1",
     "pandas>=1.4.4",
     "PyQt5>=5.15.7",
-    "dill>=0.3.8"
+    "dill>=0.3.8",
     "networkx>=3.2.1"
 ]
 dynamic = ["version"] # Version is read from uxsim/__init__.py
