### 2024-12-25 #7332 @maximelenormand
I am currently on holiday for the next three weeks. Reviewing this package in January would be complicated for me.

--
Maxime Lenormand
Research Associate at INRAE (TETIS)
Homepage: www.maximelenormand.com
Email: ***@***.***
Phone: +33 467 548 754

________________________________________
De : Kanishka Narayan ***@***.***>
Envoyé : mercredi 25 décembre 2024 01:31
À : openjournals/joss-reviews
Cc : Maxime Lenormand; Mention
Objet : Re: [openjournals/joss-reviews] [PRE REVIEW]: UXsim: lightweight mesoscopic traffic flow simulator in pure Python (Issue #7332)

@maximelenormand<https://github.com/maximelenormand>, based on your previous paper, reviews, would you be interested in reviewing this paper for JOSS? Note that this in python rather than R
