### 2024-12-28 #7332 @Nitnelav
> @Nitnelav would you be interested in reviewing this paper and package for JOSS?

Yes, I'm in. I will find some time in january.
