### 2024-12-28 #7332 @fxjung
> @fxjung would you be interested in reviewing this paper for JOSS? It seems like a good fit given previous papers youve authored

Sure, I'll also find some time in January.
