### 2024-12-28 #7617 @editorialbot
[REVIEW]: UXsim: lightweight mesoscopic traffic flow simulator in pure Python
**Submitting author:** @toruseo (<a href="http://orcid.org/0000-0001-9767-7218">Toru Seo</a>)
**Repository:** https://github.com/toruseo/UXsim
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.6.0
