### 2024-12-30 #7617 @EwoutH
@toruseo awesome to see a peer reviewed paper for UXsim. I will definitely cite it when publishing my [research](https://github.com/toruseo/UXsim/discussions/101) on self driving adoption. Thanks for the acknowledgement!
