+++ uxsim/DTAsolvers/DTAsolvers.py
@@ -41,7 +41,7 @@ class SolverDUE:
         s.W_sol = None
         s.dfs_link = []
     
-    def solve(s, max_iter, n_routes_per_od=4, swap_prob=0.05, print_info=True):
+    def solve(s, max_iter, n_routes_per_od=4, swap_prob=0.05, print_progress=True):
         """
         Solve quasi Dynamic User Equilibrium (DUE) problem using day-to-day dynamics. WIP
 
@@ -53,7 +53,7 @@ class SolverDUE:
             number of routes to enumerate for each OD pair
         swap_prob : float
             probability of route swap
-        print_info : bool
+        print_progress : bool
             whether to print the information
 
         Returns
@@ -63,7 +63,7 @@ class SolverDUE:
         """
 
         W_orig = s.func_World()
-        if print_info:
+        if print_progress:
             W_orig.print_scenario_stats()
 
         # enumerate routes for each OD pair
@@ -84,7 +84,7 @@ class SolverDUE:
             W_orig.finalize_scenario()
         dict_od_to_routes = enumerate_k_random_routes(W_orig, k=n_routes_per_od)
 
-        if print_info:
+        if print_progress:
             print(f"number of OD pairs: {len(dict_od_to_routes.keys())}, number of routes: {sum([len(val) for val in dict_od_to_routes.values()])}")
 
         # day-to-day dynamics
@@ -97,8 +97,7 @@ class SolverDUE:
         swap_prob = swap_prob
         max_iter = max_iter #50 or 100 is better
 
-        if print_info:
-            print("solving DUE...")
+        print("solving DUE...")
         for i in range(max_iter):
             W = s.func_World()
             if i != max_iter-1:
@@ -162,7 +161,7 @@ class SolverDUE:
                     n_swap += 1
                     routes_specified[key] = route_changed
 
-            if print_info:
+            if print_progress:
                 print(f' iter {i}: time gap: {total_t_gap:.1f}, potential route change: {potential_n_swap}, route change: {n_swap}, total travel time: {W.analyzer.total_travel_time: .1f}, delay ratio: {W.analyzer.average_delay/W.analyzer.average_travel_time: .3f}')
 
             s.route_log.append(route_actual)
@@ -179,6 +178,12 @@ class SolverDUE:
             #     W.analyzer.network_anim(animation_speed_inverse=15, figsize=(6,6), detailed=0, network_font_size=0, file_name="out/due_anim_1mid.gif")
             # if i == max_iter-1:
             #     W.analyzer.network_anim(animation_speed_inverse=15, figsize=(6,6), detailed=0, network_font_size=0, file_name="out/due_anim_2last.gif")
+
+        print("DUE summary:")
+        last_iters = int(max_iter/4)
+        print(f" total travel time: initial {s.ttts[0]:.1f} -> average of last {last_iters} runs {np.average(s.ttts[-last_iters:]):.1f}")
+        print(f" number of potential route changes: initial {s.potential_swaps[0]:.1f} -> average of last {last_iters} runs {np.average(s.potential_swaps[-last_iters:]):.1f}")
+        print(f" travel time gap: initial {s.total_t_gaps[0]:.1f} -> average of last {last_iters} runs {np.average(s.total_t_gaps[-last_iters:]):.1f}")
 
         s.W_sol = W
         return s.W_sol
@@ -290,7 +295,7 @@ class SolverDSO:
         s.W_sol = None
         s.dfs_link = []
     
-    def solve(s, max_iter, n_routes_per_od=4, beta_coef=100, beta_coef2=5000, print_info=True, print_info_detailed=False, plot_res=False, initial_solution_World=None):
[…]
