+++ demos_and_examples/example_26en_DUO_DUE_DSO.py
@@ -1,15 +1,21 @@
 """
-WIP
+Demonstration of dynamic traffic assginment (DTA) in a simple network.
+- Dynamic User Optimum (DUO)
+- Dynamic User Equilibrium (DUE)
+- Dynamic System Optimum (DSO)
 """
 
 from pylab import *
 import uxsim
 from uxsim.DTAsolvers import *
 
+#################################
 # scenario definition
 def create_World():
     """
     A function that returns World object with scenario informaiton. This is faster way to reuse the same scenario, as `World.copy` or `World.load_scenario` takes some computation time.
+
+    This scenario mimics a stuation where 2 pararell routes, namely highway (fast but small capacpity) and arterial road (slow but large capacity), exist between an OD pair. Travelers choose an appropriate route depending on the routing princigple.
     """
     W = uxsim.World(
         name="",
@@ -49,11 +55,10 @@ W.exec_simulation()
 W.analyzer.print_simple_stats(force_print=True)
 df_DUO = W.analyzer.basic_to_pandas()
 
-
 #################################
 # DUE
 solver_DUE = SolverDUE(create_World)
-solver_DUE.solve(max_iter=20)   # max_iter should be larger (e.g., 100). this is just for demonstration
+solver_DUE.solve(max_iter=20)   # max_iter should be larger (e.g., 100) for a better result. this is just for demonstration
 W_DUE = solver_DUE.W_sol
 W_DUE.analyzer.print_simple_stats(force_print=True)
 df_DUE = W_DUE.analyzer.basic_to_pandas()
@@ -61,7 +66,7 @@ df_DUE = W_DUE.analyzer.basic_to_pandas()
 #################################
 # DSO
 solver_DSO = SolverDSO(create_World)
-solver_DSO.solve(max_iter=20, initial_solution_World=W_DUE) # warm start up from DUE solution   # max_iter should be larger (e.g., 100). this is just for demonstration
+solver_DSO.solve(max_iter=20, initial_solution_World=W_DUE) # warm start up from DUE solution   # max_iter should be larger (e.g., 100) for a better result. this is just for demonstration
 W_DSO = solver_DSO.W_sol
 W_DSO.analyzer.print_simple_stats(force_print=True)
 df_DSO = W_DSO.analyzer.basic_to_pandas()
@@ -69,12 +74,15 @@ df_DSO = W_DSO.analyzer.basic_to_pandas()
 #################################
 # DSO by GA
 solver_DSO_GA = SolverDSO_GA(create_World)
-solver_DSO_GA.solve(max_iter=20, pop_size=20)   # max_iter should be larger (e.g., 100). this is just for demonstration
+solver_DSO_GA.solve(max_iter=20, pop_size=20)   # max_iter should be larger (e.g., 100) for a better result. this is just for demonstration
 W_DSO_GA = solver_DSO_GA.W_sol
 W_DSO_GA.analyzer.print_simple_stats(force_print=True)
 df_DSO_GA = W_DSO_GA.analyzer.basic_to_pandas()
 
-# stats
+#################################
+# Results
+
+# print stats
 print("DUO")
 print(df_DUO)
 print("DUE")

[…]
