+++ .github/workflows/measure-coverage.yml
@@ -25,7 +25,7 @@ jobs:
     - name: Install pytest other dependencies
       run: pip install pytest pytest-rerunfailures pytest-xdist pytest-cov setuptools osmnx requests
     - name: Run verifications with pytest
-      run: pytest -n auto tests/test_verification_straight_road.py tests/test_verification_route_choice.py tests/test_verification_node.py tests/test_verification_exceptional.py tests/test_verification_sioux_falls.py tests/test_verification_multilane.py tests/test_verification_taxi.py tests/test_other_functions.py --durations=0 -v --cov=uxsim --cov-report=xml --cov-config=.github/.coveragerc
+      run: pytest -n auto tests/test_verification_straight_road.py tests/test_verification_route_choice.py tests/test_verification_node.py tests/test_verification_exceptional.py tests/test_verification_sioux_falls.py tests/test_verification_multilane.py tests/test_verification_taxi.py test_verification_dta_solvers.py tests/test_other_functions.py --durations=0 -v --cov=uxsim --cov-report=xml --cov-config=.github/.coveragerc
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v5.0.2
       with:
