### 2025-01-06 7de857a
Update measure-coverage.yml
