+++ README.jp.md
@@ -26,7 +26,7 @@ UXsimは単純，軽量，柔軟であるため，研究・教育上の目的に
 - シミュレーション結果の基本的な分析と`pandas.DataFrame`，CSVへのエクスポート
 - `Matplotlib`を使用したシミュレーション結果の可視化．インタラクティブなGUIも利用可能
 - 純Pythonであることを活かした高度なカスタマイズ性．`PyTorch`を用いた深層強化学習による交通制御など，他のPythonベースのフレームワークとも直接連携可能
-- メインコードの`uxsim.py`は約1800行のコードのみ．ユーザーが簡単に理解しカスタマイズできます
+- メインコードの`uxsim.py`は約2300行のコードのみ．ユーザーが簡単に理解しカスタマイズできます
 
 ## 計算例
 
