### 2025-01-08 #7617 @Nitnelav
All right I think I'm done with reviewing this package. Every box was easily ticked as the author clearly made a reviewer's job straightforward.

This is a very well written package, respecting most of the existing best practices for a python package, including extensive unit tests and a complete documentation.

What really stands out to me is the significant effort done to provide notebooks demos and examples that explain how to use UXSim. Everything ran out of the box, and I've enjoyed learning a few things on traffic modelling along the way.

The graphs and animations are well designed and really makes analysing the simulation's output easier.

As a MATSim user I can clearly see how easier it is to use. Just a small question, in the paper you (@toruseo) mention that certain models of MATSim are the same as UXSim, could you tell me what these models are called in MATSim?

I opened two small issues but did not link them to the review issue as they are not of importance here.

Thank you @kanishkan91 for letting me review this package. I hope I've done everything according to the JOSS guidelines.
Thank you @toruseo for this very interesting work and for opening it to the scientific community.
