### 2025-01-09 ad7b896
fix `World.show_network()` file path error
