+++ uxsim/__init__.py
@@ -3,7 +3,7 @@ from .utils import *
 from .analyzer import *
 from .scenario_reader_writer import *
 
-__version__ = "1.7.1"
+__version__ = "1.7.2"
 __author__ = "Toru Seo"
 __copyright__ = "Copyright (c) 2023 Toru Seo"
 __license__ = "MIT License"
