### 2025-01-16 5ffb3fd
improve logit model computation in DSO solver
