### 2025-01-16 5d2fbe0
Add `duration_t2` arg for `W.exec_simulation()` for better timing
