### 2025-01-16 3efec5a
Fix inconsistency between `W.T` and `W.TIME`
