+++ docs/_sources/_autosummary/uxsim.Analyzer.rst.txt
@@ -4,20 +4,11 @@
 .. automodule:: uxsim.analyzer
 
    
-   
-   
-
-   
-   
    .. rubric:: Functions
 
    .. autosummary::
    
       load_font_data
-   
-   
-
-   
    
    .. rubric:: Classes
 
@@ -25,11 +16,3 @@
    
       Analyzer
    
-   
-
-   
-   
-   
-
-
-
