+++ docs/_sources/_autosummary/uxsim.uxsim.rst.txt
@@ -4,15 +4,6 @@
 .. automodule:: uxsim.uxsim
 
    
-   
-   
-
-   
-   
-   
-
-   
-   
    .. rubric:: Classes
 
    .. autosummary::
@@ -24,11 +15,3 @@
       Vehicle
       World
    
-   
-
-   
-   
-   
-
-
-
