### 2025-01-17 b056990
Minor language changes
