### 2025-01-17 #7617 @fxjung
In his manuscript *UXsim: lightweight mesoscopic traffic flow simulator in pure Python*, @toruseo introduces a traffic flow simulation library with a broad application range on the mesoscopic scale. The library is written in pure Python and easily extendable. In addition to a basic example, the manuscript also features two advanced use cases of the software: reinforcement-learning-based traffic control, and a large scale simulation using the street network of Chicago, US.

The manuscript is very well-written, making it easy to understand the purpose, functionality, and potential applications of the software package, along with its basic usage.

The software seems easy to use and fast, with a clear scope of application. I very much enjoyed the documentation, which is easy to understand, well-structured, and comprehensive.

One issue I am unable to assess is how the following existing publication (cited in the manuscript) relates to the current manuscript ([summary](https://toruseo.jp/misc/MacroTrafficSim_English_summary.pdf), [publisher page](https://www.coronasha.co.jp/np/isbn/9784339052794/)):

> Seo, T. (2023a). Macroscopic traffic flow simulation: Fundamental mathematical theory and Python implementation. Corona Publishing Co., Ltd.

I personally think it would be very worthwhile to publish the library in JOSS, making it accessible for a broad international audience. But I don't know about the exact policies of the journal. Maybe @kanishkan91 can comment on this.

Apart from that, I listed a few comments for a minor revision below.

**Software/documentation**

- The documentation doesn't specify the Python version requirements
- Automated tests:
	- The extensive test suite is run in a CI pipeline using GitHub actions, which is very nice. However, unfortunately, the documentation lacks information on how to set up the tests or generally a development environment locally. Adding this would be very helpful for future contributors.
	- The package dependencies for developing/contributing/running tests are not specified. The easiest option would probably be to specify these as `[project.optional-dependencies]` in `pyproject.toml`. These can then be installed, e.g., by `pip install uxsim[dev]`.
	- The test suite opens a large number of graphical output windows which have to be closed manually in order for the tests to finish. I think, this can be automated somehow. Or maybe I did something wrong when running the tests?
	- Having tried to install all necessary packages to run the tests locally also on a (fast) machine lacking a graphical interface, one test fails: `tests/test_result_gui_viewer.py` (test runner crashes). This may just be due to the missing graphics, but I thought I'd point it out.

**Manuscript**

- In lines 63-67, you position the library in relation to the MATSim package. From the paragraph it sounds as if UXsim's functionality is somewhat a subset of MATSim's, which sparks the question why one should chose the new library. Maybe you could add a short sentence motivating why UXsim might be preferable (possibly easier integration, usability, learning curve, leaner, ...?)
- I found it a bit confusing at first that the nodes in the code example are positioned at locations $(x,y)\in[0,2]^2$, while the edges have lengths of 1000. Maybe you could add something like "dummy locations for visualization" as a comment in the code listing.
- As the runtime estimates given in lines 132 and 147 will depend on the hardware setup, it would be great if you'd specify the reference hardware, at least roughly (standard office PC, ...).
- Deep reinforcement learning and "Chicago-Sketch" examples: I assume that these are based on the two respective tutorials in the docs? It would be great if you'd mention this in the manuscript or even directly reference the tutorials so that readers can directly jump to reproducing the examples.
- "Chicago-Sketch" example: How did you model the demand? Did you use demand data from the dataset to run the simulations?
- Fig 5
	- Link traffic volume in the left panel is missing the unit
	- The density color bars on both panels are not labeled
	- In the right panel seconds are abbreviated as "sec", compared to SI "s" in the rest of the manuscript.
- Lastly, I made some minor language edits, PR here: https://github.com/toruseo/UXsim/pull/180. I also slightly modified the code example to only import `uxsim.World` and use keyword arguments for the second call to `W.adddemand()`, too. Please feel free to merge this if you are happy with it.
