### 2025-01-24 8280d1a
Acknowledgements
equivalent to https://github.com/toruseo/UXsim/pull/168
