### 2025-01-24 828c220
Delete action from JOSS branch
