### 2025-01-24 9781fdc
Delete action from JOSS branch
