+++ .github/workflows/run-examples.yml
@@ -1,29 +1,0 @@
-name: Run examples
-
-on:
-  push:
-  pull_request:
-  workflow_dispatch:
-  schedule:
-    - cron: '0 6 * * 1'
-
-jobs:
-  run-examples:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.9", "3.12"]
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install uxsim and dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install .
-    - name: Install pytest other dependencies
-      run: pip install pytest pytest-xdist setuptools gymnasium torch osmnx deap
-    - name: Run examples with pytest
-      run: pytest -n auto tests/test_examples.py --durations=0 -v
