+++ .github/workflows/run-notebook-demos.yml
@@ -1,25 +1,0 @@
-name: Run Notebooks demos
-
-on: 
-  push:
-  pull_request:
-  workflow_dispatch:
-  schedule:
-    - cron: '0 6 * * 1'
-
-jobs:
-  run-notebook-demos:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python 3.12
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.12"
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install pytest pytest jupyter nbformat
-        pip install .
-    - name: Run tests
-      run: pytest tests/test_notebook_demos.py -s -v
