### 2025-01-24 92d4f00
Delete action from JOSS branch
