+++ .github/workflows/test-functions.yml
@@ -1,28 +1,0 @@
-name: Test functions
-
-on:
-  push:
-  pull_request:
-  workflow_dispatch:
-  schedule:
-    - cron: '0 6 * * 1'
-
-jobs:
-  test-functions:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python 3.12
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.12"
-    - name: Install uxsim and dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install .
-    - name: Install pytest other dependencies
-      run: pip install pytest pytest-rerunfailures pytest-xdist setuptools osmnx requests
-    - name: Run tests with pytest
-      run: pytest -n auto tests/test_other_functions.py --durations=0 -v
-
-
