### 2025-01-24 cffdc19
Delete action from JOSS branch
