+++ .github/workflows/test-pip.yml
@@ -1,26 +1,0 @@
-name: Test pip
-
-on:
-  release:
-    types: [published]
-  workflow_dispatch:
-  schedule:
-    - cron: '0 6 * * 1'
-
-jobs:
-  test-pip:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up Python 3.12
-      uses: actions/setup-python@v5
-      with:
-        python-version: "3.12"
-    - name: Install uxsim and dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install uxsim
-    - name: Install pytest other dependencies
-      run: pip install pytest pytest-rerunfailures pytest-xdist setuptools
-    - name: Run tests with pytest
-      run: pytest -n auto tests/test_verification_sioux_falls.py --durations=0 -v
