### 2025-01-24 e3d16b6
update tests and docs
