+++ README.md
@@ -29,6 +29,7 @@ If you are interested, please see:
 - Visualization of simulation results using `Matplotlib`; interactive GUI is also available
 - Flexible and customizable thanks to pure Python implementation; can also be directly integrated with other Python-based frameworks, such as `PyTorch` for deep reinforcement learning traffic control
 - The main code `uxsim.py` is only about 2300 lines of code. Users may easily understand and customize it
+- Python version: 3.9 or later
 
 ## Simulation Examples
 
