+++ paper/paper.bib
@@ -193,3 +193,10 @@
   doi     = {10.1007/s10109-020-00342-2},
   file    = {:Lovelace2021review.pdf:PDF},
 }
+
+@MastersThesis{terHoeven2024master,
+  author = {ter Hoeven, Ewout},
+  school = {Delft University of Technology},
+  title  = {When do autonomous vehicles solve orexacerbate different urban mobility problems?},
+  year   = {2024},
+}
