+++ paper/paper.bib
@@ -195,8 +195,8 @@
 }
 
 @MastersThesis{terHoeven2024master,
-  author = {ter Hoeven, Ewout},
+  author = {{Ter Hoeven}, Ewout},
   school = {Delft University of Technology},
-  title  = {When do autonomous vehicles solve orexacerbate different urban mobility problems?},
+  title  = {When do autonomous vehicles solve or exacerbate different urban mobility problems?},
   year   = {2024},
 }
