### 2025-01-24 #7617 @toruseo
@fxjung I have finished the revision of manuscript and package. Please see the following response letter.

----
I wish to express my appreciation to the reviewer for review and the valuable comments on my paper. They have helped us significantly improve the paper. Please find our point-by-point responses to all the comments by the reviewer below.

> In his manuscript UXsim: lightweight mesoscopic traffic flow simulator in pure Python, @toruseo introduces a traffic flow simulation library with a broad application range on the mesoscopic scale. The library is written in pure Python and easily extendable. In addition to a basic example, the manuscript also features two advanced use cases of the software: reinforcement-learning-based traffic control, and a large scale simulation using the street network of Chicago, US.
> The manuscript is very well-written, making it easy to understand the purpose, functionality, and potential applications of the software package, along with its basic usage.
> The software seems easy to use and fast, with a clear scope of application. I very much enjoyed the documentation, which is easy to understand, well-structured, and comprehensive.

> I personally think it would be very worthwhile to publish the library in JOSS, making it accessible for a broad international audience. But I don't know about the exact policies of the journal. Maybe @kanishkan91 can comment on this.

Thank you very much for the positive overall comments. I hope my library and this manuscript are useful for the communities.

> One issue I am unable to assess is how the following existing publication (cited in the manuscript) relates to the current manuscript (summary, publisher page): Seo, T. (2023a). Macroscopic traffic flow simulation: Fundamental mathematical theory and Python implementation. Corona Publishing Co., Ltd.

There is a little bit complicated background. In this book, some codes on traffic simulation are presented under MIT License with my copyright. And, UXsim is based on those codes. In addition, the book explains the mathematical models used in UXsim in very details. Thus, I cited the book in the manuscript. I understand these points are difficult to assess by the reviewer, but I believe it (i.e., citing a closed source book that is not accessible to reviewers) is acceptable in general peer-review papers.

> Apart from that, I listed a few comments for a minor revision below.

> Software/documentation

> The documentation doesn't specify the Python version requirements

It requires Python version 3.9 or later. Added this info to `README.md`.

> Automated tests: The extensive test suite is run in a CI pipeline using GitHub actions, which is very nice. However, unfortunately, the documentation lacks information on how to set up the tests or generally a development environment locally. Adding this would be very helpful for future contributors.

Added an instruction to `tests/README.md`.

> The package dependencies for developing/contributing/running tests are not specified. The easiest option would probably be to specify these as [project.optional-dependencies] in pyproject.toml. These can then be installed, e.g., by pip install uxsim[dev].

Thank you for suggestion. I agree that this is helpful for users. But, in my opinion, this approach is cumbersome to maintain. This is because the requirements are currently specified at the yaml files in `.github/workflows`, and it will be difficult to keep the consistency between these yamls and `pyproject.toml` in the root in the future updates. Thus, instead of modifying the toml file, I added instruction to `tests/README.md` to see yaml files for requirements.

> The test suite opens a large number of graphical output windows which have to be closed manually in order for the tests to finish. I think, this can be automated somehow. Or maybe I did something wrong when running the tests?

Thank you. I found that `matplotlib.use('Agg')` can automate closing operation. Implemented it.

> Having tried to install all necessary packages to run the tests locally also on a (fast) machine lacking a graphical interface, one test fails: tests/test_result_gui_viewer.py (test runner crashes). This may just be due to the missing graphics, but I thought I'd point it out.

Yes, this is an expected behavior. If machines with graphics are used, it will run without error. I had tried to make this test runnable without graphics, but I could not find a solution. Thus, I would like to keep the current implementation.

> Manuscript
> In lines 63-67, you position the library in relation to the MATSim package. From the paragraph it sounds as if UXsim's functionality is somewhat a subset of MATSim's, which sparks the question why one should chose the new library. Maybe you could add a short sentence motivating why UXsim might be preferable (possibly easier integration, usability, learning curve, leaner, ...?)

This is very true! I added the following sentence: "*UXsim* focuses on traffic flow simulation, making it simpler and easier to use and integrate with other packages".

> I found it a bit confusing at first that the nodes in the code example are positioned at locations ( x , y ) ∈ [ 0 , 2 ] 2 , while the edges have lengths of 1000. Maybe you could add something like "dummy locations for visualization" as a comment in the code listing.

I added a short sentence to the manuscript and repo's `README.md` that use a similar code.

> As the runtime estimates given in lines 132 and 147 will depend on the hardware setup, it would be great if you'd specify the reference hardware, at least roughly (standard office PC, ...).

I added this info: "on a standard office PC".

> Deep reinforcement learning and "Chicago-Sketch" examples: I assume that these are based on the two respective tutorials in the docs? It would be great if you'd mention this in the manuscript or even directly reference the tutorials so that readers can directly jump to reproducing the examples.

I added this statement: "They are also available at the GitHub repository in executable notebooks."
Since the repo and online documents may be updated in the future, I think adding direct URL to the paper may not be the best option.

> "Chicago-Sketch" example: How did you model the demand? Did you use demand data from the dataset to run the simulations?

The demand included in the official "Chicago-Sketch" dataset were used. I directed users for the online documents for the details, by adding a general explanation in the section header.

> Fig 5. Link traffic volume in the left panel is missing the unit. The density color bars on both panels are not labeled. In the right panel seconds are abbreviated as "sec", compared to SI "s" in the rest of the manuscript.

Updated the figure accordingly.

> Lastly, I made some minor language edits, PR here: Minor language changes toruseo/UXsim#180. I also slightly modified the code example to only import uxsim.World and use keyword arguments for the second call to W.adddemand(), too. Please feel free to merge this if you are happy with it.

Thank you very much for these edits. I merged the PR and made small tweaks to make it consistent to the other codes in the documentation.

Thank you very much again for the very detailed, constructive comments. I hope that the revisions in the manuscript and our accompanying responses will be sufficient to make my manuscript suitable to be published in this journal.
