### 2025-01-24 #7617 @toruseo
👆 Revised PDF

BTW,

> > Just a small question, in the paper you ([@toruseo](https://github.com/toruseo)) mention that certain models of MATSim are the same as UXSim, could you tell me what these models are called in MATSim?
>
> In my understanding, `kinematicWaves` mode of QSim in MATSim is almost the same to the UXsim's logic. Honestly speaking, my knowledge and experience on MATSim are a little bit superficial, so I will confirm the details and update the manuscript to be more specific later (maybe in the next week).

I also addressed this point. Mention to `kinematicWaves`  is added to the manuscript. @Nitnelav
