### 2025-01-24 #7617 @kanishkan91
@fxjung @toruseo - On this particular comment-

"One issue I am unable to assess is how the following existing publication (cited in the manuscript) relates to the current manuscript (summary, publisher page): Seo, T. (2023a). Macroscopic traffic flow simulation: Fundamental mathematical theory and Python implementation. Corona Publishing Co., Ltd."

The JOSS policy is to publish open source scientific software where the science does not need to be justified. It seems based on the discussion that the paper in the above comment is related to the science and equations and this paper would be about the software itself. In which case, I think its fine and we don't need to worry about double counting.
