### 2025-01-31 #7617 @fxjung
@toruseo: Thank you for the updates to the paper and for the detailed response and explanations. I only have two minor comments:

- I saw that you added the Python version requirement to the README in the "Features" section, where it's a bit hard to find if you jump right to the "Install" section as I often do. Maybe you could move it there, or, if you also want to have it above, specify it again for the installation instructions.

  (In my experience, Python versions are a common point of confusion amongst users, so I think it's important to make the minimum requirement as obvious as possible.)
- You undid the changes I made to the code listing. This is of course completely fine as this is mostly personal preference and by no means a requirement for publication. Nonetheless, I want to quickly explain my reasoning behind the changes:
  - ```python
     from uxsim import *
     # vs
     from uxsim import World
    ```
     I think that only importing `World` makes the code cleaner as it highlights that the entire simulation examples just uses a single definition from UXsim, and that no additional magic occurs somewhere hidden in the code.
  - ```python
     W.adddemand(orig="orig2", dest="dest", t_start=400, t_end=1000, flow=0.6)
     # vs
     W.adddemand("orig2", "dest", 400, 1000, flow=0.6)
    ```
     In my opinion, using keyword arguments makes the code easier to understand for the unfamiliar reader. However, my main point here was that I think the code style should be consistent in the listing, i.e., if you use keyword arguments in the first occurence of `W.adddemand`, then why not in the second? Especially since you use keyword arguments for all other calls, except for `W.addNode`.

I also found two typos which I've fixed in another PR.
