### 2025-02-03 aefb062
Add python version badge
https://github.com/openjournals/joss-reviews/issues/7617
