### 2025-02-03 d9d9241
coding style
https://github.com/openjournals/joss-reviews/issues/7617
