+++ paper/paper.md
@@ -111,7 +111,7 @@ They are also available at the GitHub repository in executable notebooks, with d
 The following code executes *UXsim* in a simple setting.
 
 ```python
-from uxsim import *
+from uxsim import World
 
 # Define the main simulation
 # Units are standardized to seconds (s) and meters (m)
@@ -119,26 +119,26 @@ W = World(
     name="",    # Scenario name
     deltan=5,   # Simulation aggregation unit delta n
     tmax=1200,  # Total simulation time (s)
-    print_mode=1, save_mode=1, show_mode=0,    # Various options
+    print_mode=1, save_mode=1, show_mode=1,    # Various options
     random_seed=0    # Set the random seed
 )
 
 # Define the scenario
 ## Create nodes
 W.addNode(name="orig1", x=0, y=0)  #xy coords are for visualization 
-W.addNode("orig2", 0, 2)
-W.addNode("merge", 1, 1)
-W.addNode("dest", 2, 1)
+W.addNode(name="orig2", x=0, y=2)
+W.addNode(name="merge", x=1, y=1)
+W.addNode(name="dest", x=2, y=1)
 ## Create links between nodes
 W.addLink(name="link1", start_node="orig1", end_node="merge",
           length=1000, free_flow_speed=20, number_of_lanes=1)
-W.addLink("link2", "orig2", "merge", 
+W.addLink(name="link2", start_node="orig2", end_node="merge", 
           length=1000, free_flow_speed=20, number_of_lanes=1)
-W.addLink("link3", "merge", "dest", 
+W.addLink(name="link3", start_node="merge", end_node="dest", 
           length=1000, free_flow_speed=20, number_of_lanes=1)
 ## Create OD traffic demand between nodes
 W.adddemand(orig="orig1", dest="dest", t_start=0, t_end=1000, flow=0.45)
-W.adddemand("orig2", "dest", 400, 1000, flow=0.6)
+W.adddemand(orig="orig2", dest="dest", t_start=400, t_end=1000, flow=0.6)
 
 # Run the simulation to the end
 W.exec_simulation()
@@ -147,9 +147,9 @@ W.exec_simulation()
 W.analyzer.print_simple_stats()
 
 # Visualize snapshots of network traffic state for several timesteps
-W.analyzer.network(100, detailed=1, network_font_size=12)
-W.analyzer.network(600, detailed=1, network_font_size=12)
-W.analyzer.network(800, detailed=1, network_font_size=12)
+W.analyzer.network(t=100, detailed=1, network_font_size=12)
+W.analyzer.network(t=600, detailed=1, network_font_size=12)
+W.analyzer.network(t=800, detailed=1, network_font_size=12)
 ```
 
 The meaning of each function would be intuitive for Python users.
