### 2025-02-03 #7617 @fxjung
@toruseo Thanks for the updates and reconsidering the style changes to the code listing. I think the code example really highlights how easy it is to set up and run a simulation using UXsim, which is great.

Let me thank you again for creating and publishing this very cool package, and taking the time to submit it to JOSS!
From my side, the manuscript is good to go!

@kanishkan91 I recommend accepting the manuscript without further revision.
