### 2025-02-03 #7617 @kanishkan91
👋 @toruseo - Looks like we are getting close! I'm taking a read through the latest manuscript. In the meanwhile, we need to set up the archive for your new release.

So as next steps for you:

- [ ] Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.

- [ ] Check that software archive exists and title and author list match JOSS paper (or purposefully do not).

- [ ] Please respond with the 1) version and 2) DOI of the archived version here
