### 2025-02-04 #7617 @toruseo
@kanishkan91 Thanks for the instruction. I have executed those steps, and the software is archived in Zenodo as follows.

1) version: v1.6.0
2) DOI: https://doi.org/10.5281/zenodo.14799163
