### 2025-02-04 #7617 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
