### 2025-02-06 #7617 @toruseo
@crvernon I archived v1.6.0 because it was the version in the submission information. Versions v1.6.0 and v.1.7.2 are not very different, especially regarding the content of the paper. So I think v1.6.0 is fine.

If I overlook something, please let me know.
