+++ README.md
@@ -234,6 +234,10 @@ When publishing works based on UXsim, please cite:
 }
 ```
 
+If you need more detailed information, please also cite:
+- Toru Seo. [Macroscopic Traffic Flow Simulation: Fundamental Mathematical Theory and Python Implementation](https://toruseo.github.io/misc/MacroTrafficSim_English_summary.pdf). Corona Publishing Co., Ltd., 2023.
+- Toru Seo. [UXsim: An open source macroscopic and mesoscopic traffic simulator in Python-a technical overview](http://dx.doi.org/10.48550/arXiv.2309.17114). arXiv preprint arXiv: 2309.17114, 2023.
+
 Works using UXsim is summarized on the [Github Wiki page](https://github.com/toruseo/UXsim/wiki). Please feel free to edit.
 
 ## Contributing and Discussion
@@ -258,5 +262,3 @@ Specifically, UXsim directly uses the following works:
 
 - [Toru Seo (Author)](https://toruseo.jp/index_en.html)
 - [Seo Laboratory, Tokyo Institute of Technology](http://seo.cv.ens.titech.ac.jp/en/)
-- Toru Seo. [Macroscopic Traffic Flow Simulation: Fundamental Mathematical Theory and Python Implementation](https://toruseo.github.io/misc/MacroTrafficSim_English_summary.pdf). Corona Publishing Co., Ltd., 2023.
-- Toru Seo. [UXsim: An open source macroscopic and mesoscopic traffic simulator in Python-a technical overview](http://dx.doi.org/10.48550/arXiv.2309.17114). arXiv preprint arXiv: 2309.17114, 2023
