### 2025-02-07 225c178
Delete .github/CITATION.cff
