### 2025-02-07 d54d732
Create CITATION.cff
